split login component and container
